[youbot gazebo control] use urdf::JointConstSharedPtr
With ROS Kinetic, C++11 is finally officially in REP-3
http://www.ros.org/reps/rep-0003.html

The urdf::JointConstSharedPtr is a alias for std::shared_ptr on kinetic and newer.
https://github.com/ros/urdfdom_headers

But for Indigo compatibility headers are available.
https://github.com/ros/urdf/tree/indigo-devel

This PR should compile with Indigo, Kinetic and Melodic. (tested on Melodic)

More could be done, all of the boost::shared_ptr and boost usage in general could be replaced.